update: 구조 일부 수정
